Fix diffusion bug in cosmic code; add tools for testing cosmic ray strategies on simulated image cubes.

+++ cosmical_realistic/cosmical.h
@@ -1,1 +1,1 @@
-double *cosmical(double crfl, double exptm1, double exptm2, long NX, long NY, long idodif);
+double *cosmical(double crfl, double exptm1, double exptm2, long NX, long NY, int idodif);
